requirements.txt
langchain 
google-generativeai 
langchain-google-genai 
python-dotenv 
streamlit
st_btn_select